agents/requirements.txt
# CrewAI Scraping Agents (using Claude)
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
playwright>=1.40.0

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.0

# Supabase
supabase>=2.3.0

# Environment
python-dotenv>=1.0.0

# Rate limiting
ratelimit>=2.2.1
